feat: restructure template
