+++ src/collections/index.ts
@@ -16,7 +16,6 @@ import { LinkCollection } from './domain/useCases/LinkCollection'
 import { UnlinkCollection } from './domain/useCases/UnlinkCollection'
 import { GetCollectionLinks } from './domain/useCases/GetCollectionLinks'
 import { GetCollectionsForLinking } from './domain/useCases/GetCollectionsForLinking'
-import { CreateDatasetTemplate } from './domain/useCases/CreateDatasetTemplate'
 
 const collectionsRepository = new CollectionsRepository()
 
@@ -37,7 +36,6 @@ const linkCollection = new LinkCollection(collectionsRepository)
 const unlinkCollection = new UnlinkCollection(collectionsRepository)
 const getCollectionLinks = new GetCollectionLinks(collectionsRepository)
 const getCollectionsForLinking = new GetCollectionsForLinking(collectionsRepository)
-const createDatasetTemplate = new CreateDatasetTemplate(collectionsRepository)
 
 export {
   getCollection,
@@ -56,8 +54,7 @@ export {
   linkCollection,
   unlinkCollection,
   getCollectionLinks,
-  getCollectionsForLinking,
-  createDatasetTemplate
+  getCollectionsForLinking
 }
 export { Collection, CollectionInputLevel } from './domain/models/Collection'
 export { CollectionFacet } from './domain/models/CollectionFacet'
