+++ src/datasets/infra/repositories/DatasetsRepository.ts
@@ -1,4 +1,3 @@
-import { AxiosResponse } from 'axios'
 import { ApiRepository } from '../../../core/infra/repositories/ApiRepository'
 import { IDatasetsRepository } from '../../domain/repositories/IDatasetsRepository'
 import { Dataset, VersionUpdateType } from '../../domain/models/Dataset'
@@ -25,9 +24,6 @@ import { DatasetLinkedCollection } from '../../domain/models/DatasetLinkedCollec
 import { CitationFormat } from '../../domain/models/CitationFormat'
 import { transformDatasetLinkedCollectionsResponseToDatasetLinkedCollection } from './transformers/datasetLinkedCollectionsTransformers'
 import { FormattedCitation } from '../../domain/models/FormattedCitation'
-import { DatasetTemplate } from '../../domain/models/DatasetTemplate'
-import { DatasetTemplatePayload } from './transformers/DatasetTemplatePayload'
-import { transformDatasetTemplatePayloadToDatasetTemplate } from './transformers/datasetTemplateTransformers'
 import { DatasetType } from '../../domain/models/DatasetType'
 import { TermsOfAccess } from '../../domain/models/Dataset'
 import { transformTermsOfAccessToUpdatePayload } from './transformers/termsOfAccessTransformers'
@@ -397,18 +393,6 @@ export class DatasetsRepository extends ApiRepository implements IDatasetsReposi
       true
     )
       .then((response) => response.data.data as string[])
-      .catch((error) => {
-        throw error
-      })
-  }
-
-  public async getDatasetTemplates(
-    collectionIdOrAlias: number | string
-  ): Promise<DatasetTemplate[]> {
-    return this.doGet(`/dataverses/${collectionIdOrAlias}/templates`, true)
-      .then((response: AxiosResponse<{ data: DatasetTemplatePayload[] }>) =>
-        transformDatasetTemplatePayloadToDatasetTemplate(response.data.data)
-      )
       .catch((error) => {
         throw error
       })
