+++ src/index.ts
@@ -12,3 +12,4 @@ export * from './notifications'
 export * from './search'
 export * from './licenses'
 export * from './externalTools'
+export * from './template'
