+++ src/template/domain/repositories/ITemplatesRepository.ts
@@ -1,0 +1,10 @@
+import { CreateDatasetTemplateDTO } from '../dtos/CreateDatasetTemplateDTO'
+import { DatasetTemplate } from '../models/DatasetTemplate'
+
+export interface ITemplatesRepository {
+  createDatasetTemplate(
+    collectionIdOrAlias: number | string,
+    template: CreateDatasetTemplateDTO
+  ): Promise<void>
+  getDatasetTemplates(collectionIdOrAlias: number | string): Promise<DatasetTemplate[]>
+}
